traffiCtimeTable now  uses generators
Changed the hierarchy of the generator. TrafficTimeTabke has generators not the other way around
Bug fixed: TrafficLaw and TrafficPeriod had the same type

+++ src/ummisco/gama/traffgen/generators/AgentSeed.java
@@ -33,7 +33,7 @@ class AgentSeed {
 	public double getActivationDate() {
 		return activationDate;
 	}
-	public void setActivationDate(float activationDate) {
+	public void setActivationDate(double activationDate) {
 		this.activationDate = activationDate;
 	}
 	public ILocation getStartingPoint() {
